pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["EUNJU HWANG <dev.owov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]  
python = "^3.11"
django = "^5.1.1"
pillow = "^10.4.0"
djangorestframework = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"   
